refactoring, fixes, removing storage json file on cluster removal
